+++ CHANGELOG.md
@@ -35,6 +35,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - The internal function `path_resolution()` returns `ELOOP` if too many
   symbolic links were followed
 - The internal function `path_resolution()` follows relative symlinks
+- The internal function `path_resolution()` does not ignore `/run/systemd`
 
 ### Fixed